pyproject.toml
[tool.poetry]
name = "project-1"
version = "0.1.0"
description = ""
authors = ["Abhishek Sharma <abhishek.sharma@uplight.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pymysql = "^1.1.1"
mysqlclient = "^2.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
